add hidden config command to manage lab configuration in a KV store
Rather than serve lab config from a static file and it's own service,
push the details into a KV store; in this case consul.  There should
be enough here to manage all the machines in a fine-grained manner -
adjusting single elements of a COM port config for example.

Should be able to load from the existing lab.map to make the
transition easier.